+++ components/ui/ApiReference/MultiLangExample/MultiLangExample.tsx
@@ -4,7 +4,7 @@ import { useEffect, useMemo } from "react";
 import { useIsMounted } from "../../../../hooks/useIsMounted";
 import useLocalStorage from "../../../../hooks/useLocalStorage";
 import { CodeBlock, SupportedLanguage } from "../../CodeBlock";
-import { EVENT_NAME, LOCAL_STORAGE_KEY } from "../../../MultiLangCodeBlock";
+import { EVENT_NAME, LOCAL_STORAGE_KEY } from "../../MultiLangCodeBlock";
 
 type Props = {
   examples: Record<string, string>;
