+++ components/ui/LocaleTable.tsx
