+++ content/cli.mdx
@@ -424,7 +424,8 @@ Note:
 <Attributes>
   <Attribute
     name="--environment"
-    type="string (required)"
+    type="string"
+    isRequired
     description="The environment to activate the workflow in."
   />
   <Attribute
