+++ content/getting-started/quick-start.mdx
@@ -72,6 +72,7 @@ In this guide, you'll integrate Knock with your backend web application and send
     <MultiLangCodeBlock
       snippet="workflows.trigger-with-user-identification"
       title="Trigger your workflow"
+      mb="2"
     />
 
     Learn more about trigger calls in our [API reference](/reference#trigger-workflow).
