+++ pages/[...slug].tsx
@@ -2,13 +2,11 @@ import fs from "fs";
 import { join, sep } from "path";
 import { serialize } from "next-mdx-remote/serialize";
 import { MDXRemote } from "next-mdx-remote";
-import Image from "next/image";
 import rehypeAutolinkHeadings from "rehype-autolink-headings";
 import remarkGfm from "remark-gfm";
 import remarkSlug from "remark-slug";
 import rehypeMdxCodeProps from "rehype-mdx-code-props";
 
-import { CodeBlock } from "../components/ui/CodeBlock";
 import MDXLayout from "../layouts/MDXLayout";
 import {
   getAllFilesInDir,
@@ -17,61 +15,13 @@ import {
   makeIdFromPath,
   generateAlgoliaIndex,
 } from "../lib/content.server";
-import Callout from "../components/Callout";
-import MultiLangCodeBlock from "../components/MultiLangCodeBlock";
 import eventPayload from "../data/code/sources/eventPayload";
-import LocaleTable from "../components/LocaleTable";
-import { Attributes, Attribute } from "../components/Attributes";
-import Table from "../components/Table";
-import CopyableText from "../components/CopyableText";
-import { Card, CardGroup } from "../components/Card";
-import { SdkCard, SdkCardGroup } from "../components/SdkCard";
-import { Endpoints, Endpoint } from "../components/ui/Endpoints";
-import { AccordionGroup, Accordion } from "../components/Accordion";
-import { Steps, Step } from "../components/Step";
 import datadogDashboardJson from "../content/integrations/extensions/datadog_dashboard.json";
 import newRelicDashboardJson from "../content/integrations/extensions/new_relic_dashboard.json";
 import AiChatButton from "../components/AiChatButton";
 
-import {
-  Section,
-  ContentColumn,
-  ErrorExample,
-  ExampleColumn,
-} from "../components/ui/ApiSections";
-import RateLimit from "../components/RateLimit";
 import { FrontMatter } from "../types";
-import { Heading } from "@telegraph/typography";
-
-export const MDX_COMPONENTS = {
-  pre: CodeBlock,
-  h2: (props) => <Heading as="h2" size="5" mb="3" {...props} />,
-  h3: (props) => <Heading as="h3" size="4" mb="2" {...props} />,
-  h4: (props) => <Heading as="h4" size="3" mb="1" {...props} />,
-  Callout,
-  Image,
-  MultiLangCodeBlock,
-  LocaleTable,
-  Attributes,
-  Attribute,
-  Table,
-  CopyableText,
-  RateLimit,
-  Section,
-  ContentColumn,
-  ErrorExample,
-  ExampleColumn,
-  Endpoints,
-  Endpoint,
-  Card,
-  CardGroup,
-  SdkCard,
-  SdkCardGroup,
-  AccordionGroup,
-  Accordion,
-  Steps,
-  Step,
-};
+import { MDX_COMPONENTS } from "@/lib/mdxComponents";
 
 export default function ContentPage({ source, sourcePath }) {
   return (
