+++ pages/mapi-reference/index.tsx
@@ -7,7 +7,7 @@ import remarkGfm from "remark-gfm";
 import { readOpenApiSpec, readStainlessSpec } from "../../lib/openApiSpec";
 import ApiReference from "../../components/ui/ApiReference/ApiReference";
 import { CONTENT_DIR } from "../../lib/content.server";
-import { MDX_COMPONENTS } from "../[...slug]";
+import { MDX_COMPONENTS } from "@/lib/mdxComponents";
 import { SidebarSection } from "../../data/types";
 
 export const RESOURCE_ORDER = [
