+++ styles/global.css
@@ -15,12 +15,6 @@
   margin-top: 0;
 }
 
-.tgraph-content h2,
-.tgraph-content h3,
-.tgraph-content h4 {
-  margin-top: var(--tgph-spacing-8);
-}
-
 .tgraph-content p, 
 .tgraph-content ul, 
 .tgraph-content ol {
@@ -36,16 +30,6 @@
 .tgraph-content p, li {
   font-size: var(--tgph-text-2);
   line-height: var(--tgph-leading-3);
-}
-
-.tgraph-content li {
-  margin-bottom: var(--tgph-spacing-1);
-}
-
-.tgraph-content p code {
-  font-family: var(--tgph-family-mono);
-  font-size: var(--tgph-text-1);
-  color: var(--tgph-accent-9);
 }
 
 .tgraph-content p a,
@@ -77,6 +61,7 @@
   color: var(--tgph-gray-12);
 }
 
+/* Data properties are new */
 [data-property-row-description] p {
   font-size: 13px;
   margin-bottom: 0;
@@ -85,4 +70,26 @@
 
 [data-property-row-container]:last-child {
   border-bottom: none;
+}
+
+.tgraph-content h2 [data-tgph-code] {
+  font-size: 18px;
+  font-weight: var(--tgph-font-weight-semi-bold);
+}
+
+.tgraph-content h3 [data-tgph-code] {
+  font-size: 16px;
+  font-weight: var(--tgph-font-weight-semi-bold);
+}
+
+.tgraph-content p [data-tgph-code] {
+  font-size: 13px;
+}
+
+.tgraph-content a [data-tgph-code] {
+  color: var(--tgph-accent-9);
+}
+
+.tgraph-content [data-example-column] [data-code-block] {
+  margin-bottom: 0;
 }
